pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "governance-os"
version = "0.2.0"
description = "A governance runtime for pipeline contract management — discover, validate, and report on structured pipeline contracts in any repository."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = [
    "governance",
    "pipeline",
    "contract",
    "validation",
    "devops",
    "cli",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "markdown-it-py>=3.0.0",
    "networkx>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
"Source" = "https://github.com/ramjf/governance-os"
"Bug Tracker" = "https://github.com/ramjf/governance-os/issues"

[project.scripts]
govos = "governance_os.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/governance_os"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------

[tool.coverage.run]
source = ["src/governance_os"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line length — handled by black
    "SIM108", # ternary — not always clearer
    "B008",  # typer.Option/Argument in function defaults is standard Typer usage
]

[tool.ruff.lint.isort]
known-first-party = ["governance_os"]

# ---------------------------------------------------------------------------
# Black
# ---------------------------------------------------------------------------

[tool.black]
line-length = 100
target-version = ["py311"]
